Now handles Google Calendar alert emails directly...
Previous versions used to rely on http://iftt.com to send an email when a new calendar event was detected.  I updated the IMAP fetch code to handle multipart messages (which is what is sent by Google) and updated the main application to expect the Google Calendar alert email format.